requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-arm64
aiosignal=1.2.0=pypi_0
attrs=21.4.0=pypi_0
bzip2=1.0.8=h3422bc3_4
c-ares=1.18.1=h3422bc3_0
ca-certificates=2022.5.18.1=h4653dfc_0
certifi=2022.5.18.1=pypi_0
cffi=1.15.0=py39h52b1de0_0
charset-normalizer=2.0.12=pypi_0
click=8.1.3=pypi_0
cloudpickle=2.1.0=pypi_0
distlib=0.3.4=pypi_0
docker-pycreds=0.4.0=pypi_0
filelock=3.7.0=pypi_0
freetype=2.10.4=h17b34a0_1
frozenlist=1.3.0=pypi_0
giflib=5.2.1=h27ca646_2
gitdb=4.0.9=pypi_0
gitpython=3.1.27=pypi_0
grpcio=1.46.3=py39h365d37b_0
idna=3.3=pypi_0
jpeg=9e=h1c322ee_1
jsonschema=4.5.1=pypi_0
lcms2=2.12=had6a04f_0
lerc=3.0=hbdafb3b_0
libblas=3.9.0=14_osxarm64_openblas
libcblas=3.9.0=14_osxarm64_openblas
libcxx=14.0.3=h6a5c8ee_0
libdeflate=1.10=h3422bc3_0
libffi=3.4.2=h3422bc3_5
libgfortran=5.0.0.dev0=11_0_1_hf114ba7_23
libgfortran5=11.0.1.dev0=hf114ba7_23
liblapack=3.9.0=14_osxarm64_openblas
libopenblas=0.3.20=openmp_h2209c59_0
libpng=1.6.37=hf7e6567_2
libprotobuf=3.20.1=h332123e_0
libtiff=4.3.0=h2810ee2_4
libwebp=1.2.2=h0d20362_0
libwebp-base=1.2.2=h3422bc3_1
libxcb=1.13=h9b22ae9_1004
libzlib=1.2.11=h90dfc92_1014
llvm-openmp=14.0.3=hd125106_0
lz4-c=1.9.3=hbdafb3b_1
msgpack=1.0.3=pypi_0
ncurses=6.3=h07bb92c_1
ninja=1.11.0=hf86a087_0
numpy=1.22.4=py39h7df2422_0
openjpeg=2.4.0=h062765e_1
openssl=3.0.3=ha287fd2_0
pathtools=0.1.2=pypi_0
pillow=9.1.1=py39h1b8be2f_0
pip=22.1.1=pyhd8ed1ab_0
platformdirs=2.5.2=pypi_0
promise=2.3=pypi_0
psutil=5.9.1=pypi_0
pthread-stubs=0.4=h27ca646_1001
pycparser=2.21=pyhd8ed1ab_0
pyrsistent=0.18.1=pypi_0
python=3.9.12=h14b404e_1_cpython
python-dateutil=2.8.2=pypi_0
python_abi=3.9=2_cp39
pytorch=1.11.0=cpu_py39h03f923b_1
pyyaml=6.0=pypi_0
ray=1.12.1=pypi_0
readline=8.1=hedafd6a_0
requests=2.27.1=pypi_0
sentry-sdk=1.5.12=pypi_0
setproctitle=1.2.3=pypi_0
setuptools=62.3.2=py39h2804cbe_0
shortuuid=1.0.9=pypi_0
six=1.16.0=pyh6c4a22f_0
sleef=3.5.1=h156473d_2
smmap=5.0.0=pypi_0
sqlite=3.38.5=h40dfcc0_0
tk=8.6.12=he1e0b03_0
torchvision=0.2.2=py_3
tqdm=4.64.0=pypi_0
typing_extensions=4.2.0=pyha770c72_1
tzdata=2022a=h191b570_0
urllib3=1.26.9=pypi_0
virtualenv=20.14.1=pypi_0
wandb=0.12.16=pypi_0
wheel=0.37.1=pyhd8ed1ab_0
xorg-libxau=1.0.9=h27ca646_0
xorg-libxdmcp=1.1.3=h27ca646_0
xz=5.2.5=h642e427_1
zlib=1.2.11=h90dfc92_1014
zstd=1.5.2=hd705a24_1
